minor refactors and renamed for multiple projectiles per shot PR
