+++ crates/thetawave_interface/src/character.rs
@@ -35,8 +35,8 @@ pub struct Character {
     pub projectile_direction: f32,
     /// Time until fired projectile despawns
     pub projectile_despawn_time: f32,
-    /// Velocity of fired projectile
-    pub projectile_velocity: f32,
+    /// Base speed of fired projectile
+    pub projectile_speed: f32,
     /// Position of projectile spawn relative to player
     pub projectile_offset_position: Vec2,
     /// Period of time between firing blasts
