+++ src/assets/consumable.rs
@@ -28,13 +28,8 @@ impl ConsumableAssets {
         }
     }
 
+    #[allow(unused)] // Placeholder for if we put this in the item config files
     pub fn get_color(&self, consumable_type: &ConsumableType) -> Color {
-        match consumable_type {
-            ConsumableType::Money1 => Color::rgb(1.6, 1.6, 1.6),
-            ConsumableType::Money3 => Color::rgb(1.6, 1.6, 1.6),
-            ConsumableType::HealthWrench => Color::rgb(1.6, 1.6, 1.6),
-            ConsumableType::Armor => Color::rgb(1.6, 1.6, 1.6),
-            ConsumableType::GainProjectiles => Color::rgb(1.6, 1.6, 1.6),
-        }
+        Color::rgb(1.6, 1.6, 1.6)
     }
 }
