+++ src/main.rs
@@ -2,21 +2,11 @@ use bevy::app::PluginGroupBuilder;
 use bevy::{asset::AssetPlugin, pbr::AmbientLight, prelude::*};
 use bevy_kira_audio::prelude::*;
 
-use bevy_rapier2d::geometry::Group;
 use bevy_rapier2d::prelude::*;
 use options::{generate_config_files, GameInitCLIOptions};
 use thetawave_interface::states::{AppStates, GameStates};
 
 pub const PHYSICS_SCALE: f32 = 10.0;
-
-// Collider groups used for rapier physics
-pub const SPAWNABLE_COLLIDER_GROUP: Group = Group::GROUP_1;
-pub const HORIZONTAL_BARRIER_COLLIDER_GROUP: Group = Group::GROUP_2;
-pub const VERTICAL_BARRIER_COLLIDER_GROUP: Group = Group::GROUP_3;
-pub const ALLY_PROJECTILE_COLLIDER_GROUP: Group = Group::GROUP_4;
-pub const ENEMY_PROJECTILE_COLLIDER_GROUP: Group = Group::GROUP_5;
-pub const NEUTRAL_PROJECTILE_COLLIDER_GROUP: Group = Group::GROUP_6;
-pub const MOB_COLLIDER_GROUP: Group = Group::GROUP_7;
 
 mod animation;
 mod arena;
